ai/app/requirements.txt
annotated-types @ file:///C:/b/abs_0dmaoyhhj3/croot/annotated-types_1709542968311/work
anyio @ file:///C:/b/abs_847uobe7ea/croot/anyio_1706220224037/work
asttokens==2.4.1
Bottleneck @ file:///C:/b/abs_f7un855idq/croot/bottleneck_1709069969633/work
certifi @ file:///C:/b/abs_35d7n66oz9/croot/certifi_1707229248467/work/certifi
click @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/click_1699473800648/work
colorama @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/colorama_1699472650914/work
comm==0.2.1
debugpy==1.8.1
decorator==5.1.1
distro @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/distro_1701796812765/work
executing==2.0.1
fastapi @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/fastapi_1707431081657/work
h11 @ file:///C:/b/abs_1czwoyexjf/croot/h11_1706652332846/work
httpcore @ file:///C:/b/abs_55n7g233bw/croot/httpcore_1706728507241/work
httpx @ file:///C:/b/abs_70wr6jzznw/croot/httpx_1706887145357/work
idna @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/idna_1699473483982/work
ipykernel==6.29.3
ipython==8.22.1
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.1
matplotlib-inline==0.1.6
mkl-fft @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/mkl_fft_1699473528480/work
mkl-random @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/mkl_random_1699473588250/work
mkl-service==2.4.0
nest-asyncio==1.6.0
numexpr @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/numexpr_1699503421264/work
numpy @ file:///C:/b/abs_c1ywpu18ar/croot/numpy_and_numpy_base_1708638681471/work/dist/numpy-1.26.4-cp312-cp312-win_amd64.whl#sha256=becc06674317799ad0165a939a7613809d0bee9bd328a1e4308c57c39cacf08c
openai @ file:///C:/b/abs_130pqu0rxo/croot/openai_1706892467657/work
pandas @ file:///C:/b/abs_3awk0iw2ab/croot/pandas_1709590545218/work/dist/pandas-2.2.1-cp312-cp312-win_amd64.whl#sha256=8f9837b9f672189b7e2df7dcf64b91243a78b0fad6e1125220b33a39b5c9e598
pandas-stubs @ file:///C:/b/abs_e0sp1h4xvt/croot/pandas-stubs_1705009933752/work
parso==0.8.3
platformdirs==4.2.0
prompt-toolkit==3.0.43
pure-eval==0.2.2
pydantic @ file:///C:/b/abs_f6ffn86wjx/croot/pydantic_1709578044014/work
pydantic_core @ file:///C:/b/abs_05smrkbwjm/croot/pydantic-core_1709574000211/work
python-dateutil==2.9.0.post0
python-dotenv @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/python-dotenv_1699475097728/work
pytz @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pytz_1699475334900/work
pywin32==306
pyzmq==25.1.2
setuptools==68.2.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio @ file:///C:/b/abs_3akdewudo_/croot/sniffio_1705431337396/work
stack-data==0.6.3
starlette @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/starlette-recipe_1699544517909/work
tornado==6.4
tqdm @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/tqdm_1701808178601/work
traitlets==5.14.1
types-pytz @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/types-pytz_1699545516418/work
typing_extensions @ file:///C:/b/abs_23uoivfoy8/croot/typing_extensions_1705619982210/work
tzdata @ file:///croot/python-tzdata_1690578112552/work
uvicorn @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/uvicorn-split_1699568993341/work
wcwidth==0.2.13
wheel==0.41.2
